+++ randomizer-core/src/test/java/de/alpharogroup/random/date/RandomDateExtensionsTest.java
@@ -34,6 +34,10 @@ import static org.testng.AssertJUnit.assertTrue;
 import java.lang.reflect.InvocationTargetException;
 import java.sql.Timestamp;
 import java.text.ParseException;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.ZoneId;
 import java.util.Date;
 
 import org.meanbean.test.BeanTestException;
@@ -84,7 +88,7 @@ public class RandomDateExtensionsTest extends BaseTestCase
 	}
 
 	/**
-	 * Test method for {@link RandomDateExtensions#randomBirthday()}.
+	 * Test method for {@link RandomDateExtensions#randomBirthday()}
 	 */
 	@Test
 	public void testCreateRandomBirthday()
@@ -123,11 +127,14 @@ public class RandomDateExtensionsTest extends BaseTestCase
 	@Test
 	public void testCreateRandomDate()
 	{
-		final Date randomDate = RandomDateExtensions.randomDate();
-		assertNotNull(randomDate);
+		for (int i = 0; i < 7; i++)
+		{
+			final Date randomDate = RandomDateExtensions.randomDate();
+			assertNotNull(randomDate);
 
-		actual = !randomDate.equals(this.now);
-		assertTrue(actual);
+			actual = !randomDate.equals(this.now);
+			assertTrue(actual);
+		}
 	}
 
 	/**
@@ -293,6 +300,52 @@ public class RandomDateExtensionsTest extends BaseTestCase
 	{
 		Timestamp timestamp = RandomDateExtensions.getTimestamp(now);
 		assertNotNull(timestamp);
+	}
+
+	/**
+	 * Test method for {@link RandomDateExtensions#randomLocalDate()}
+	 */
+	@Test
+	public void testRandomLocalDate()
+	{
+		for (int i = 0; i < 7; i++)
+		{
+			LocalDate randomObject = RandomDateExtensions.randomLocalDate();
+			assertNotNull(randomObject);
+		}
+	}
+
+	/**
+	 * Test method for {@link RandomDateExtensions#randomLocalDateTime()}
+	 */
+	@Test
+	public void testRandomLocalDateTime()
+	{
+		LocalDateTime randomObject = RandomDateExtensions.randomLocalDateTime();
+		assertNotNull(randomObject);
+	}
+
+	/**
+	 * Test method for {@link RandomDateExtensions#randomLocalTime()}
+	 */
+	@Test
+	public void testRandomLocalTime()
+	{
+		for (int i = 0; i < 7; i++)
+		{
+			LocalTime randomObject = RandomDateExtensions.randomLocalTime();
+			assertNotNull(randomObject);
+		}
+	}
+
+	/**
+	 * Test method for {@link RandomDateExtensions#randomZoneId()}
+	 */
+	@Test
+	public void testRandomZoneId()
+	{
+		ZoneId randomObject = RandomDateExtensions.randomZoneId();
+		assertNotNull(randomObject);
 	}
 
 	/**